pyproject.toml
[project]
name = "graphql_api"
dynamic = ["version"]
description = "A framework for building Python GraphQL APIs."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Robert Parker", email = "rob@parob.com"}
]
keywords = ["GraphQL", "GraphQL-API", "GraphQLAPI", "Server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "graphql-core",
    "requests",
    "typing-inspect",
    "aiohttp",
    "docstring-parser",
    "pydantic",
]

[project.urls]
Homepage = "https://gitlab.com/parob/graphql-api"
Repository = "https://gitlab.com/parob/graphql-api"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "coverage",
    "faker",
    "flake8",
    "flake8-pyproject",
    "sphinx",
    "sphinxcontrib-fulltoc",
    "guzzle_sphinx_theme",
    "autopep8>=2.3.2",
]

[tool.uv]
package = false

[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["graphql_api"]
include-package-data = true

[tool.setuptools.package-data]
graphql_api = [
    "introspect.graphql"
]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:PYTEST_DONT_REWRITE:DeprecationWarning",
    "ignore:ast.Str is deprecated and will be removed in Python 3.14; use ast.Constant instead:DeprecationWarning:docstring_parser.attrdoc"
]

[tool.flake8]
max-line-length = 170
ignore = [
    "E203",  # whitespace before ':' (conflicts with black)
    "W503",  # line break before binary operator (conflicts with black)
    "W504",  # line break after binary operator (conflicts with W503)
    "E402",  # module level import not at top of file (sometimes needed)
]

